test(recipe-scraper): run the pyodide bundle tests in CI
The suite asserted on modules/recipe-scraper/assets/pyodide-bundle.html and
self-skipped when that file was absent:

  const describeIfBundle = bundleExists ? describe : describe.skip;

The bundle is gitignored and only produced by setup-pyodide.sh during an iOS
prebuild, while the CI unit job is just setup-node + test:unit:coverage. So
these 16 tests had never run in CI, and coverage reported green regardless.

Every assertion is about template text and ordering, not the multi-megabyte
payloads, so the artifact was never actually needed. Extract the pure document
builder out of generate-pyodide-html.mjs into buildBundleHtml.js and assert on
its return value with stub payloads. The generator keeps all its IO and calls
the builder; the template was moved byte-for-byte, so the emitted bundle is
unchanged.

Adds one assertion that the supplied payloads reach the output, since stubs
would otherwise satisfy the existing checks trivially.

Trade-off: the asset-reading half of the generator (missing wheels, bad
Pyodide files) is no longer covered by these tests — it was never covered by
them in CI anyway — and remains verified by an actual iOS prebuild.

Unit suite goes from 211 passed / 1 skipped to 212 passed / 0 skipped, and
4248 passed / 16 skipped to 4265 passed / 0 skipped.
